+++ CHANGES.md
@@ -8,6 +8,7 @@
   * Theme switcher no longer changes theme instantly bug follows Apply/Cancel pattern.
   * Separated user theme configuration pane from Appearance config.
   * Replaced use of Timer with hour aligned data source to make widged less resource hungry.
+  * Theme editor features "Fake calendar" mode now, to make theme testing a breeze.
   * Fixed widget tooltip's sub-text not working properly.
 
 * v1.4.0 (2021-01-10)